Moved parse_message from ingester to helpers
