+++ lambdas/ingester.py
@@ -1,7 +1,7 @@
 import re
 import json
 import os
-from botocore.vendored import requests  # TODO: Use the requests library.
+import requests
 import helpers
 
 # False when setup has not been performed.
@@ -26,7 +26,7 @@ def setup():
     humio_repository = os.environ['humio_repository_name']
     humio_ingest_token = os.environ['humio_ingest_token']
 
-    http_session = requests.session()
+    http_session = requests.Session()
 
     global _is_setup
     _is_setup = True
@@ -51,8 +51,8 @@ def lambda_handler(event, context):
     # TODO: Use Python Client.
     humio_url = '%s://%s/api/v1/dataspaces/%s/ingest' % (humio_protocol, humio_host, humio_repository)
     humio_headers = {
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer %s' % humio_ingest_token
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer %s' % humio_ingest_token
     }
 
     # Decode and unzip the log data.
@@ -89,7 +89,7 @@ def lambda_handler(event, context):
         # Create the attributes.
         attributes = {}
         attributes.update(batch_attrs)
-        attributes.update(parse_message(message))
+        attributes.update(helpers.parse_message(message))
 
         # Append the flattened event
         humio_events.append({
@@ -105,7 +105,7 @@ def lambda_handler(event, context):
     # Make request.
     request = http_session.post(
         humio_url,
-        data=json.dumps(wrapped_data).encode(),  # encode might not be necessary.
+        data=json.dumps(wrapped_data),
         headers=humio_headers
     )
 
@@ -113,65 +113,3 @@ def lambda_handler(event, context):
 
     # Debug output.
     print('Got response %s from Humio.' % response)
-
-
-# Standard out from Lambdas.
-std_matcher = re.compile(
-    '\d\d\d\d-\d\d-\d\d\S+\s+(?P<request_id>\S+)'
-)
-
-
-# END RequestId: b3be449c-8bd7-11e7-bb30-4f271af95c46
-end_matcher = re.compile(
-    'END RequestId:\s+(?P<request_id>\S+)'
-)
-
-
-# START RequestId: b3be449c-8bd7-11e7-bb30-4f271af95c46
-# Version: $LATEST
-start_matcher = re.compile(
-    'START RequestId:\s+(?P<request_id>\S+)\s+'
-    'Version: (?P<version>\S+)'
-)
-
-
-# REPORT RequestId: b3be449c-8bd7-11e7-bb30-4f271af95c46
-# Duration: 0.47 ms
-# Billed Duration: 100 ms
-# Memory Size: 128 MB
-# Max Memory Used: 20 MB
-report_matcher = re.compile(
-    'REPORT RequestId:\s+(?P<request_id>\S+)\s+'
-    'Duration: (?P<duration>\S+) ms\s+'
-    'Billed Duration: (?P<billed_duration>\S+) ms\s+'
-    'Memory Size: (?P<memory_size>\S+) MB\s+'
-    'Max Memory Used: (?P<max_memory>\S+) MB'
-)
-
-
-def parse_message(message):
-    """
-    Simple CloudWatch Logs parser.
-
-    :param message: Log event message.
-    :type message: str
-
-    :return: Parsed message or empty.
-    :rtype: dict
-    """
-    m = None
-
-    # Determine which matcher to use depending on the message type.
-    if message.startswith('END'):
-        m = end_matcher.match(message)
-    elif message.startswith('START'):
-        m = start_matcher.match(message)
-    elif message.startswith('REPORT'):
-        m = report_matcher.match(message)
-    else:
-        m = std_matcher.match(message)
-
-    if m:
-        return m.groupdict()
-    else:
-        return {}
